Trajectory redesign (#106)
* simplify traces

* remove get_trace name

* fix tests

* implement isfull

* deprecate get_trace

* resolve comments

* do not show internals of environments

* use Flux@v0.11.1

* automatic copy view/reshaped-view to GPU without copying parent

+++ src/ReinforcementLearningCore/src/utils/sum_tree.jl
@@ -63,7 +63,7 @@ mutable struct SumTree{T} <: AbstractArray{Int,1}
     length::Int
     nparents::Int
     tree::Vector{T}
-    SumTree(capacity::Int) = SumTree(Float64, capacity)
+    SumTree(capacity::Int) = SumTree(Float32, capacity)
     function SumTree(T, capacity)
         nparents = 2^ceil(Int, log2(capacity)) - 1
         new{T}(capacity, 1, 0, nparents, zeros(T, nparents + capacity))
